+++ lib/image_util/codec.rb
@@ -62,7 +62,7 @@ module ImageUtil
       end
 
       def detect_io(io)
-        Magic.detect_io(io)
+        Magic.detect_io(io).first
       end
 
       private
@@ -105,7 +105,6 @@ module ImageUtil
     autoload :Pam, "image_util/codec/pam"
     autoload :ImageMagick, "image_util/codec/image_magick"
     autoload :RubySixel, "image_util/codec/ruby_sixel"
-    autoload :Magic, "image_util/codec/magic"
 
     register_codec :Pam, :pam
     register_codec :Libpng, :png